move EDI to it's own file
